malé úpravy vzhľadu a odstránenie starých chybových hlášok

+++ App/Views/Markers/index.view.php
@@ -34,30 +34,30 @@ if ($auth->isLogged()) { ?>
 
             let element;
             <?php foreach ($data as $marker) { ?>
-                element = document.createElement('div');
-                element.className = 'marker';
-                element.style.backgroundImage = `url(public/images/markers/marker-<?= $marker->getColor(); ?>.png)`;
+            element = document.createElement('div');
+            element.className = 'marker';
+            element.style.backgroundImage = `url(public/images/markers/marker-<?= $marker->getColor(); ?>.png)`;
 
-                new mapboxgl.Marker(element, {anchor: 'bottom'})
-                    .setLngLat(['<?= $marker->getLong(); ?>', '<?= $marker->getLat(); ?>'])
-                    .setPopup(
-                        new mapboxgl.Popup({offset: 25})
-                            .setHTML(
-                                `<h4><?= $marker->getTitle(); ?></h4>` +
-                                <?php if ($marker->getPhoto()) { ?>
-                                `<img src="<?= $marker->getPhoto() ?>" class="card-img-top" alt="...">` +
-                                <?php } ?>
-                                `<p class="mt-2" ><?= $marker->getDescription(); ?></p>` +
-                                `<div class="row">` +
-                                `<p id="marker-rating" class="col"><span data-m-id="<?= $marker->getId(); ?>" class="rating-number"><?= $marker->getRating(); ?></span> <span class="rating-star">★</span></p>` +
-                                `</div>`
-                                <?php if ($auth->isLogged() && $marker->getAuthorId() == $auth->getLoggedUserId()) { ?>
-                                + `<a href="?c=markers&a=edit&id=<?= $marker->getId(); ?>" class="btn btn-warning m-1">Upraviť</a>` +
-                                `<a href="?c=markers&a=delete&id=<?= $marker->getId(); ?>" class="btn btn-danger m-1">Zmazať</a>`
-                                <?php } ?>
-                            )
-                    )
-                    .addTo(map);
+            new mapboxgl.Marker(element, {anchor: 'bottom'})
+                .setLngLat(['<?= $marker->getLong(); ?>', '<?= $marker->getLat(); ?>'])
+                .setPopup(
+                    new mapboxgl.Popup({offset: 25})
+                        .setHTML(
+                            `<h4><?= $marker->getTitle(); ?></h4>` +
+                            <?php if ($marker->getPhoto()) { ?>
+                            `<img src="<?= $marker->getPhoto() ?>" class="card-img-top" alt="...">` +
+                            <?php } ?>
+                            `<p class="mt-2" ><?= $marker->getDescription(); ?></p>` +
+                            `<div class="row">` +
+                            `<p id="marker-rating" class="col"><span data-m-id="<?= $marker->getId(); ?>" class="rating-number"><?= $marker->getRating(); ?></span> <span class="rating-star">★</span></p>` +
+                            `</div>`
+                            <?php if ($auth->isLogged() && $marker->getAuthorId() == $auth->getLoggedUserId()) { ?>
+                            + `<a href="?c=markers&a=edit&id=<?= $marker->getId(); ?>" class="btn btn-warning m-1">Upraviť</a>` +
+                            `<a href="?c=markers&a=delete&id=<?= $marker->getId(); ?>" class="btn btn-danger m-1" onclick="return confirm('Naozaj chcete odstrániť túto lokalitu?');">Zmazať</a>`
+                            <?php } ?>
+                        )
+                )
+                .addTo(map);
             <?php } ?>
             map.on('style.load', function () {
                 map.on('dblclick', function (e) {
